refactor: separate component for alignment controls
